fix(user-management): improve relative URL management and add error messages
affects: @zetapush/user-management

When the user provides a relative URL in configuration, it is concatenated to the URL that is
automatically provided by ZetaPush according to run environment.
In previous version, some URLs were
not totally compliant with this system.
Moreover, there are some cases where the URLs adaptated to
the run environment can't be automatically provided (for example in local when the front is not
started). In this case, an explicit error indicates that the front must be started to benefit from
automatic URLs or to explicitly provide an absolute URL.

BREAKING CHANGE:
'registration.confirmation.base-url' and 'reset-password.ask.base-url' have been removed. Please use
'registration.confirmation.confirm-url' and 'registration.confirmation.confirm-url' instead